+++ argos/controllers/__init__.py
@@ -1,5 +1,6 @@
 from argos.controllers.albums import AlbumsController
 from argos.controllers.base import ControllerBase
+from argos.controllers.images import ImagesController
 from argos.controllers.library import LibraryController
 from argos.controllers.playback import PlaybackController
 from argos.controllers.playlists import PlaylistsController
@@ -9,6 +10,7 @@ from argos.controllers.volume import MixerController
 __all__ = (
     "AlbumsController",
     "ControllerBase",
+    "ImagesController",
     "LibraryController",
     "PlaybackController",
     "PlaylistsController",